fix: split API client into public and admin modules (#1158)
The full admin API surface (resolveMarket, emailPreview, emailSendTest,
getEmailAnalytics, getEmailQueueStats) and the entire Soroban contract
error-code map (CONTRACT_ERROR_MESSAGES, ~64 entries) were bundled into
the public landing-page build because client.ts was imported wholesale
by Statistics.tsx and LandingPage.tsx.

Changes:
- public-client.ts: new module containing only public-facing endpoints
  (health, getStatistics, getFeaturedMarkets, getContent, blockchain
  read-only, newsletter self-service). No admin methods, no contract
  error map.
- admin-client.ts: new module that re-exports the public api and adds
  admin-only endpoints + CONTRACT_ERROR_MESSAGES/getContractErrorMessage.
- client.ts: replaced with a thin re-export barrel from admin-client.ts
  so all existing imports continue to work without changes.
- Statistics.tsx, LandingPage.tsx: updated to import from public-client.ts
  so the landing-page bundle no longer pulls in admin code.
- public-client.test.ts: 17 new tests asserting admin symbols are absent
  from the public module and present on the admin module.

All 78 tests pass. next build succeeds with zero warnings.
